Move run_transaction into transaction.rs
